backend/itmanagement/requirements.txt
asgiref==3.9.1
attrs==25.3.0
cachetools==5.5.2
certifi==2025.7.14
cffi==1.17.1
channels==4.2.2
charset-normalizer==3.4.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
cryptography==45.0.5
decorator==5.2.1
dj-database-url==3.0.1
Django==5.2.4
django-axes==8.0.0
django-channels==0.7.0
django-cors-headers==4.7.0
django-otp==1.6.1
djangorestframework==3.16.0
djangorestframework_simplejwt==5.5.0
drf-spectacular==0.28.0
drf-yasg==1.21.10
google-auth==2.40.3
idna==3.10
imageio==2.37.0
imageio-ffmpeg==0.6.0
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.6
joblib==1.5.1
jsonschema==4.24.1
jsonschema-specifications==2025.4.1
MarkupSafe==3.0.2
moviepy==2.2.1
numpy==2.2.6
oauthlib==3.3.1
opencv-python==4.12.0.88
packaging==25.0
pillow==11.3.0
proglog==0.1.12
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.22
PyJWT==2.9.0
python-decouple==3.8
python-dotenv==1.1.1
pytz==2025.2
PyYAML==6.0.2
referencing==0.36.2
requests==2.32.4
requests-oauthlib==2.0.0
rpds-py==0.26.0
rsa==4.9.1
ruamel.yaml==0.18.14
ruamel.yaml.clib==0.2.12
scikit-learn==1.7.1
scipy==1.16.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
threadpoolctl==3.6.0
tqdm==4.67.1
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
whitenoise==6.9.0
